Fix build on windows, need explicit cast
git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/projects/FatCRM/trunk@36983 c66179b2-ff0d-0410-8aef-1b1a91e65cfa

+++ client/src/dbuswinidprovider.cpp
@@ -69,7 +69,7 @@ DBusWinIdProvider::~DBusWinIdProvider()
 
 qlonglong DBusWinIdProvider::getWinId()
 {
-    return d->mWindow->winId();
+    return (qlonglong) d->mWindow->winId();
 }
 
 #include "dbuswinidprovider.moc"
